move getCredit function to the only place it is now used
